requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
pydantic[email]
bcrypt==4.0.1
aiofiles